examples/mcp-chatbot/requirements.txt
llama-stack-client==0.3.3

python-dotenv

httpx
typing-extensions
